FluentWinUI3: Improve Dialog when using high contrast themes
The Dialog type was lacking the iconic outline which Dialogs get when
high contrast themes are enabled on Windows 11.

Add it to both Dialog, and DialogButtonBox, and make adjustments in
order to make a Dialog with and without a DialogButtonBox look similar
to a Windows App SDK ContentDialog, when using a high contrast theme.

Task-number: QTBUG-129088
Change-Id: Ie158a46566dceaa343f9120cc44969b410b5abe4
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 23df452196c165ee8fbc7a4efc82a357bca5ca16)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>