Popup/Overlay: Filter no events for non-popup children of Overlay
Previously in 01de98703916ef485f8934485270478b90a524f6 we allowed
MousePress and MouseRelease events to be passed to non-popup children of
the overlay, but no other pointer event types.

Allowing just MousePress and MouseRelease events seems arbitrary, since
we should either tell users that they can put items on top of the
overlay, or discourage this kind of use case altogether.

Since 81e53a7e9cf7821c4e512c40671f56ec9b0c52fc we documented that this
is a practice that should work normally, and thus it would make sense to
allow all types of events to be sent to items that are on top of the
overlay.

Fixes: QTBUG-132644
Task-number: QTBUG-134545
Pick-to: 6.9 6.8 6.5
Change-Id: Iad29b9abb694bd0653da220f2922b294219396bb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>