Doc: add a section about mixing style selection methods
We don't want to encourage this, but as mentioned in the change
itself: in the case of applications that load third party QML code, it
may not be possible to control which imports are used. So, this change
documents the limitations to be aware of when mixing the two.

Task-number: QTBUG-117526
Task-number: QTBUG-136709
Pick-to: 6.9 6.8 6.5
Change-Id: I5601266758f123998d3ae5d03892fba42c6c5b8f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit a6548a92512cfc976ef828d220a9429c62e6cc36)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>