doc: Discourage users from abusing the overlay in unintended ways
Upon further investigation, it's clear that the overlay only exist in
order to make popup's closePolicy and modality possible.

Unfortunately, due to the various event filtering that happens in the
overlay, stacking items on top of the overlay can sometimes cause the
overlay to filter events that should be sent to items that are above it.

Because of this, we shouldn't advice users to abuse the overlay in ways
that it was never intended to be used. However, since there are no other
solution to displaying items in front of the overlay, it might be
equally problematic to remove the section entirely.

Lets at least discourage users from doing this, while we come up with a
more streamlined solution.

Task-number: QTBUG-134545
Pick-to: 6.9 6.8 6.5
Change-Id: I2a20824996c6bdffd4b0fb76ebd8111c08effd06
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>