qmllint: Do not trigger assert on typo when using Math
The code handling accesse to members of the math object asserted that
there's exactly one method of a given name. However, that is only true
if a valid function is called. If there's a typo (or some other mistake,
or someone tried to extend the math object), we will find 0 matches.

Detect this case, and print a warning.

Pick-to: 6.9
Change-Id: Iaf3f445662660940d8b3d77e3a1d40fc39c6f2a4
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>