Make GridView::reuseItems work
Amends 941b36a98646ae9361e0e72ffebf60bd8cb856d4.

pooled() and reused() signals had been integrated but the items are not
actually reused because we also need to tell the QQuickItemView that an
item can be reused when it is released.

This change was tested in Plasma Wayland, it makes scrolling wallpapers
butter smooth.

Fixes: QTBUG-101209
Pick-to: 6.9
Change-Id: Id7f99e0dc90a8fdff1f8cd137d7a52d40f8cf70f
Reviewed-by: Marco Martin <mart@kde.org>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>