Popup window: disable window "fitting" on Wayland
Wayland surfaces don't know where they are on the screen, which means
that we can't rely on QWindow::mapToGlobal({0,0}) to get their position.

It looks like we're supposed to use the wayland api to position popups.
More specifically a xdg_positioner.

Disable the part that doesn't work, until we figure out a proper
solution. QTBUG-99618 tracks necessary changes that neeeds to be made to
qtbase and the wayland platform window, before we can fix it in
qtdeclarative.

Fixes: QTBUG-135158
Pick-to: 6.9 6.9.1 6.8
Change-Id: I9af0250f77bddc0c4f09bb028cd5105d4548a061
Reviewed-by: Liang Qi <liang.qi@qt.io>