qmlls: silence file watcher warning
addPaths() warns when called on an empty list, so avoid calling it when
its argument is empty.

Note: A QML module that does not have any C++ types will not refer to
any C++ headers in its .qmltypes file. In this case no files paths are
to be watched.

Task-number: QTBUG-133497
Change-Id: I12143093066f8ef080a17b0aa1914ca5efca0962
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>