a11y: reverse the mapping of Quick to QAccessibleInterface relationships
The Quick properties are declaring the relationship from the perspective
of the current element: a text element is the label for an input field.

The QAccessibleInterface::relations API is for finding a related object:
return the object that is labelled by the current object.

Amends 644349fecb9ce58659e590be7049ba93889d9f34.

Change-Id: I6f6138fe45969a95a270e185da3e272bb3b7dfc7
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
Reviewed-by: Nicolas Fella <nicolas.fella@kdab.com>