sg: Fix dpr when grabbing a not-yet-shown QQuickWindow
The scenegraph is not initialized when the window has not been
exposed yet, and calling QQuickWindow::grabWindow() takes a whole
different code path in this case. This was not communicating a
scale factor to the renderer at all.

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-116675
Change-Id: I73b21dfd06a04dbdc2ff3cb1a17bbe0364ca0624
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit d96f616722fc7b3f49ae671ecd20f50517920a96)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>