qmlls: remove unused indexing code
It seems that we have some code in qmlls that runs an "indexing" thread
in the background to load non-opened files. It is not tested and never
used : the only commit using the indexing functionality is
2717985373dc3e2765ced402c775a8b4386b1e93 but never sets `m_status` to
`Index`, so that the indexing is almost never run.

There are some code paths, for example when a
DidChangeWorkspaceFoldersNotification is received by the server, where
the indexing is triggered despite it being disabled.

Enabling the indexing leads to a ASAN crash, probably because of some
missing synchronization on some shared data like:
* QQmlJSImporter and QQmlJSResourceFileMapper: both needed to construct
  the Dom while loading a QML file,
* QQmlJScope::Ptr: the same ptr might be lazyloaded by two different
  threads at the same time)

Remove the indexing funtionality from the code, and pick the change
back to avoid memory errors on DidChangeWorkspaceFoldersNotifications.

This could be related to one of the qmlls crashes described in the
comments of QTBUG-133586.

Task-number: QTBUG-133586
Pick-to: 6.9 6.8
Change-Id: I547cf11150dc25b04208a83bf5e123e7ee5b0482
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit c77ff3ffc4c24d2d6bb899c8772278e2a6edfdf1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>