cmake/qt_add_qml_module: Fix QtQuick auto dependency setup
We need to actually defer the code which attempts to detect whether we
are linking against QtQuick to the point where we write-out the file, as
the user is not unlikely to use target_link_libraries _after_
qt_add_qml_module.
Amends 065b784cab3ff8e8a1dd1f5a304c6b8c3f222d79.

To test that this works, we run the all_qmllint target of on an example
project. We need to add a .qmllint.ini file to let qmllint fail the
build if there are any warnings.

Pick-to: 6.9 6.8
Change-Id: I7e00583b3a5b38549db2be13b1a8817b8489496d
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 0051e751c3773f63de5259bf9377cab7c6068422)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>