Qml: Warn about using enums as type annotations in functions
The warning emitted when calling functions with parameters type
annotated with enums does not communicate what's wrong or how to fix
it:

"1 should be coerced to void because the function called is
insufficiently annotated. The original value is retained. This will
change in a future version of Qt."

Until we support a proper representation of enums in the type system,
all we can do is warn the user that this does not work.

Add another warning that fires once when the function is constructed
that has a better description of what is going wrong. This should
provide the user with the info to solve the issue unlike the original
warning.

This does mean that we warn more for the same issue but it should still
be beneficial overall.

Task-number: QTBUG-135255
Pick-to: 6.9 6.8 6.5
Change-Id: Icd8f8d80c6a4b9bd2b33c660394b9cf5a228a346
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit f64ba5dec875927163bb917597bac51540eb4097)
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>