QQuickMultiPointTouchArea: fix grab threshold calculation
Ths logic of deciding if we need to grab a gesture was using
scenePosition() and scenePressPosition() and comparing the diff to the
styleHints()->startDragDistance().

If the scene is scaled, the returned points are also scaled, but the
startDragDistance() value is not affected by the scale. As a result,
the logic for grabbing the gesture was broken, starting it too early
(if scale > 1.0), or too late (if scale < 1.0).

Fix it by calling mapFromScene() on both points.

Fixes: QTBUG-131886
Pick-to: 6.9 6.8 6.5
Change-Id: Ie7c7472f9d9806d18728b4059e23ec279090db6f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 52086f74e6d4211d66462e5377fc909b97be4523)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>