Dom: Prefer writeRegion over write where possible in the OutWriter
As the name implies writeRegion takes the region into account. This
helps keep the comments in the right place. If we write the content of a
region manually, any comments that are associated with it may be left
behind and removed from the formatted code.

Additionally, use ensureSpace instead of baking the space into the
written string to avoid superfluous space insertions.

This change was not applied everywhere as we don't have region types
for every use case of write. They will either have to be added to cover
the remaining entries of be deemed out of scope. One such case is the
use of .pragma directive at the top of a JS file.

Task-number: QTBUG-133315
Task-number: QTBUG-123386
Pick-to: 6.9 6.8
Change-Id: I30e0b5ec354e46fdcef18b068618a0368555bd41
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
(cherry picked from commit bc6b4940a1da48ca1f82f9ccc9b9901a6c52bf56)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>