iOS Style: Resolve UIColor-s with appropriate trait collection
QQuickIOSSTheme::intialize() is called when first setting up the theme
(and its palette) for the iOS style, as well as whenever the color
scheme gets updated. This is a function that is called on the QML
thread by the style plugin, as the QQuickTheme also lives in the QML
thread.
In this function we resolve the palette colors through the dynamic
UIColor-s, which update based on the currentTraitCollection.
However, it is not safe to rely on the value of the
currentTraitCollection property outside the main thread, or even in
functions other than the documented methods where UIKit sets the
property as it may be invalid.
To fix, construct a trait collection based on the current
color scheme and resolve the UIColor against it.

Fixes: QTBUG-138200
Pick-to: 6.9 6.8
Change-Id: I404428844bc2e3b35be5746746d8d21bde3b3832
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 4eee74ecf71b1ed3c1b37683790c46fbe0d73ac3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>