FluentWinUI3 style: Don't set imlicitWidth for combobox content item
We were setting the implicitWidth of the combobox's inner text field to
depend on the width of the content (ie: text). However, for an
editable combobox this means that the textField will expand
horizontally as you type to accomodate for the new typed text. This
is not wanted behavior.
To fix, don't manually set the implicitWidth to the content item,
as a text field already has its own implicit size. Note that
"implicitResize", which controls whether the textfield
recalculates its implicit size when its content changes, is set to
false by default for a text field.

Pick-to: 6.9 6.8
Change-Id: I2f7ca2706dc2957fd600a0c06dc28010bfb4f7c6
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit d136963fd6c121c7904717ce6e3ab6d9d5011c7e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>