Doc: Clarify stacking behavior in QQuickitem documentation
The documentation of QQuickitem::stackBefore() and stackAfter()
incorrectly stated that the sibling item was being moved. The
updated documentation now correctly states that the current
item (this) is moved before or after the sibling in the
stacking order.

Pick-to: 6.9
Fixes: QTBUG-77201
Change-Id: I608f0eb5bc5338e58ba82ab656ea7eb1d83e2e28
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 0554999c0634108cf8742e0b16bce41ede6d8cec)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>