qmltyperegistrar: Do not process duplicate metatypes.json files
If we get the same file twice, that's generally the effect of a linkage
cycle. Either we're dealing with a static build where this is fine and
we don't need to complain about it, or the linker will complain anyway
and we don't have to.

Pick-to: 6.9 6.8
Fixes: QTBUG-132518
Fixes: QTBUG-127133
Fixes: QTBUG-134292
Change-Id: Iad8b3b5c694d96ebdcbf03522ce2318ae8d24fb2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 5d8a3c1367103a6704910432a6051445e3f01640)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>