FluentWinUI3: Improve the ProgressBar when using high contrast themes
The ProgressBar would continue to use the same background, even when a
contrast theme was applied. Not only did this cause the background to
have the wrong colors, but it also lacked the outline that the native
WinUI3 ProgressBar has.

Fix it by using a regular Rectangle as the background item in case
a high contrast theme is being used.

In addition, we reduce the inset by 1 pixel, in order to make room for
the new outline, without affecting the size of the contentItem.

Task-number: QTBUG-129088
Change-Id: I9b67d89d86f7d7d556c0461e34e2f23e4fb0a0be
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit ff3589bda13b1e74502d26d7ba163bd95403532e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>