Use Accessible.name for StaticText in QAccessibleQuickItem
Currently it is not possible to select QAccessible::StaticText
with the common keyboard shortcuts of screen readers on Windows.
QWindowsUiaTextProvider uses QAccessibleTextInterface but
QAccessibleQuickItem does not provide text in
text(QAccessible::Value) because QAccessible::Value is logically
only implemented for QAccessible::EditableText. So we need to
provide QAccessible::Name in the QAccessibleTextInterface.

Fixes: QTBUG-137860
Pick-to: 6.9 6.8
Change-Id: If5da452985b838d3b4c1dff2444fdea2618aaaf6
Reviewed-by: Shawn Rutledge 🏖️ <shawn.rutledge@qt.io>
(cherry picked from commit 30453147f7770e1fad34735240cbd9283814c683)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>