qmllint: Avoid confusing duplicate property assignment warning
If we have no clue what a certain property might be, do not warn about
duplicate binding assignments to it.
It might be a genuine error, but it might also be:
- a property of a type we couldn't resolve; that should warn about the
  type, but not here – it might after all be a list property if we were
  able to resolve the type
- a binding assignment to an instance of a custom parsed type

Note that the attached test case currently would still complain about
the property being unknown, as it fails to correctly handle the grouped
binding. That will be addressed in a follow-up commit; for now we
silence that additional warning with a qmllint comment, putting the
focus on the duplicate list warning.

Task-number: QTBUG-138164
Change-Id: Ib06a4baaef0813a45a20738cbb6efe4cf5e5952c
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit d7288cbe97e29ad6caa5d31245a72b6b08160406)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>