QQmlJSImportVisitor: Avoid faulty unknown property warning
In QQmlJSImportVisitor::processPropertyBindingObjects, we can encounter
grouped properties. So far, we used the whole compount expression to
look up the property. That would of course not find any property.
We now instead traverse the individual parts of the name, and look up
the property chain correctly on the corresponding types.

We adjust the previously added test for unknown duplicate properties to
work with _really_ unknown properties (that don't exist), and instead
use an adjusted version to prove that the example is completely warning
free – mirroring the issue from QTBUG-138164 in a more minimal way.

Task-number: QTBUG-138498
Task-number: QTBUG-138164
Change-Id: I75d17c3c412159f3f6da082f54d99bdf6b3b08e4
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit c7eaffdea665df1678e11cdf4b512d489fa52378)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>