FluentWinUI3: Resolve palette against platform theme and app's palette when updating
When we were updating the theme as a result of the color scheme change,
we were just updating the palette to the default colors, forgetting to
resolve against the theme palette and the guiApp palette. This was
different from the palette initialization logic, which resolves against
both platform and app palette, making the style palette inconsistent
between the initial palette and the updated one after the color scheme
change. To fix, use the same logic when both initializing and updating
the palette. Also, refactor some code and function names.

Pick-to: 6.9 6.8
Change-Id: I80a0983509323002e8d422b79c5271c312d8d963
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit da96a15dfd0e29537d32c367ed066f8bff52958d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>