macOS style: Give Switch an outline when "Increase contrast" is enabled
The Switch doesn't use any elements from the QuickNativeStyle module
and thus won't automatically get an outlined pixmap to use as a
background. In other words, we have to create the outline ourselves.

Luckily, in 6.10, we have the necessary API to do just that.

To make the Switch's look and feel be as similar to a native macOS
switch as possible, we'll have to give both the background and the
handle an outline. Since Rectangle renders its border inwards, I had to
wrap the SwitchHandle's root Rectangle in a new, slightly larger
Rectangle, which is normally invisible, unless "Increase contrast" is
enabled. This was done, so that the borders for both the SwitchIndicator
and the SwitchHandle will overlap, without causing the outline to look
twice as thick.

Change-Id: Idc1f02a230894ccce4ce211ef9499c5e729d651c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 1f95eef1e48b30e6ff3bef0da03a99cfd112eb61)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>