spreadsheets example: Configure qmllint
The example is now warning free, so we can configure qmllint to fail if
there is any additional warning.
This is unfortunately not enforced in CI yet, as the all_qmllint target
is (intentionally) not part of the all target; so we don't run qmllint
in CI so far.

Change-Id: Ib179364ac5bda033ad6a2c3d87bc3738a98d1489
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit ef84bfb5e8becf5a89840c6e6247333fe415718d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>