PointerHandler: Add handler to parent also on componentComplete()
It was possible for a handler to not be added to its parent item via
data_append() when the handler is created as a property binding:
```
Item {
    id: parentItem
    property HoverHandler handler: HoverHandler {
        parent: parentItem
    }
}
```
In that case, the HoverHandler won't be added to the parent item's
default list property (data), since it's being assigned to a different
property instead (handler).

data_append() was the main way that pointer handlers installed
themselves to items, but it is skipped in this case.

Now we also call QQuickItemPrivate::addPointerHandler() in
componentComplete() if the pointer handler was not already added.

Change-Id: I5c797f6abcfb19af7e897354fba39dd536e66140
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 51c5bc0122b54e7199c0e75825425e7079202643)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>