Remove support of manipulation of complex rows in QQmlTableModel
TableModel supports complex row structures but only with a limited
functionality. Since the model does not know how a complex row is
structured, it cannot manipulate it. This also means that the copy of
the model data that is stored in rows is not kept in sync with the
source data that was set in QML - the user is supposed to handle
simple data manipulation.

This one functionality was not tested and it seems that the
implementation was changed in the background as setRows is now
called when rows is modified in QML.

A good example is in complex.qml. If we try to extend the
corresponding test case with a setData call it is going to call the
function assigned to setDisplay which calls setRows - something it
was not supposed to do. The checks in setRows do not recognize the
complex row structure and as such the call is not successful.

The checks could be adjusted to work for this specific case, but
the structure could be anything so that solution would not be
generic enough.

Remove the support of manipulation of complex row structures
from QQmlTableModel.

[ChangeLog][Qt labs QML Models] Removed the support of manipulation
of complex row structures from QQmlTableModel.

Task-number: QTBUG-138553
Change-Id: I8664ae9ec2e5cd09bc2c88a25f4ac972378e8e94
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 705c71b5f69a8a2211c94618b3750c8a7733882b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>