Remove support of manipulation of complex rows in QQmlTreeModel
TreeModel uses the same QQmlTableModelColumns as QQmlTableModel and it
provides the same support to complex row structures.

The manipulation of these complex row structures is no longer working
as it was intended in the TableModel and the support is removed - do
the same for the TreeModel.

Task-number: QTBUG-138553
Change-Id: I3eab04f918f2e9dc689beef51457413eacfbb7ec
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3f2c325c7c676e070cf30f336379993a4b9f0a3f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>