Model/View documentation: remove link to outdated YouTube video
The video is based on Qt 5.9, uses deprecated features such as context
properties, and after 7 years could generally need an overhaul.

The video is still there on YouTube, but we don't need to direct people
to a 45 minutes tutorial that isn't up-to-date.

Change-Id: Ifb473c44fab7544f797223841cbb6823a97f7c1b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 57984600584cb7968ea9940e15880a4b2b7221f5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>