Curve renderer: Fix filling of certain paths with degenerate parts
Determining the fill side of a segment would fail if the path
contained an overlapping segment in the opposite direction. If a
subpath started with such a segment, the resulting fill could be
incorrect. Fix by iterating over the subpath for a segment with a
determinable fill side.

Fixes: QTBUG-133247
Pick-to: 6.9 6.8
Change-Id: I07c4f884d525b19af93609bcb31b412d340bebd5
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit cc161f0d35e590c3269151c2c4f3643050805aa6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>