Reorder the functions in QmlTreeModel
Nothing is changed in the functions. They are only reordered to match the
order in QmlTableModel. That allows to see differences better when
looking at the diff.

Functions are reordered in QmlTreeModel because the commit history is
not yet very meaningful.

This is the 3rd patch in an attempt to make QQmlTreeModel and
QQmlTableModel more similar and eventually merge common parts
in a common abstract class.

Task-number: QTBUG-138703
Change-Id: Idc7f7c0fedb0143b1764d614dcd8191ea68a74e1
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 2f06f3bafa76407a2e82e1852bc00c3f2f586f8c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>