Derive QmlTableModel from QAbstractItemModel
Previously QmlTableModel was derived from QAbstractTableModel. The
upsides of QAbstractTableModel are very limited but disable the
possibility to combine the similar aspects of QmlTableModel and
QmlTreeModel in a common class.

In QQuickHeaderView, QAbstractTableModel is treated in a special way.
In Order to keep all existing code working, we are treating
QmlTableModel the exact same way also after this change.

This is the 6th patch in an attempt to make QQmlTreeModel and
QQmlTableModel more similar and eventually merge common parts
in a common abstract class.

Task-number: QTBUG-138703
Change-Id: Ifb1f12d147e659320bddccf6c183ac0308cfb99b
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 6ce91b03de361836a6263937ccf086e34b9dc4d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>