qmllint: Do not warn about BINDABLE only properties (just now)
While there is a goal to phase out BINDABLE only properties in QML, and
require NOTIFY (if they participate in a binding), we are not there yet.

Given that this is not documented outside of wiki.qt.io, and we still
had Qt internal properties without NOTIFY (but with BINDABLE), we
shouldn't warn about it right now.

In the future, we should provide a more specific warning; which then can
also add more details about the integration of signals with QProperty.

Change-Id: I5070c47124787938ef3ba5143a02574663de1cf6
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit b114eedf82a048b6eef24a74ce1a95b9ec321c84)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>