Fix qt.qpa.menus warning in tst_qquickmenu
The warning was:

QWARN: tst_QQuickMenu::FluentWinUI3::nativeStatic() qt.qpa.menus: QPlatformMenuItem(0x60c0001c5040) does not belong to <QCocoaNSMenu: 0x6070000884a0>
        Title:
        Supermenu: 0x0 (None), autoenable: YES
        Previous menu: 0x0 (None)
        Next menu: 0x0 (None)
        Items:     (
    )

In nativeStatic.qml, we have an item tree that looks like this:

    Menu {
        Action { /* ... */ }

        MenuItem {
            action: Action { /* ... */ }
        }

        Menu {
            Action { /* ... */ }
        }
    }

Removing the first two menu items of the top-level menu works. When we
get to the menu item for subMenu, we see that it has a Menu of its own,
and so we call recursivelyDestroyNativeSubMenus. That removes all of
the menu's native items, and at the end it does this:

    menuPrivate->resetNativeData();

Here, menuPrivate refers to subMenu's private object. This line severs
the connection between subMenu and menu, which leads to the warning. We
then return from recursivelyDestroyNativeSubMenus and end up here in
QQuickMenuPrivate::removeNativeItem:

    if (nativeItem->handle()) {
        handle->removeMenuItem(nativeItem->handle());
        syncWithNativeMenu();
    }

The removeMenuItem call then triggers the warning.

To fix this, we need to have control over when resetNativeData is
called, so that we can do it after the removeMenuItem call. Do so by
rewriting removeNativeItem to be recursive, and using it in favour of
recursivelyDestroyNativeSubMenus, which we then remove.

Improve internal docs while we're at it.

Pick-to: 6.9 6.8
Change-Id: I4955a5e4d8e5aa5c9d46873b8b3a27f65446ea35
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 202fef1694df0c249e62489e1264a2902825975e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>