Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: Icebe063937cdf1e2ef034b723cce646cd6c96bbe
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit f695b0bd5dd9b281ee15e7cd595d0c5786796836)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>